+++ src/Container.hs
@@ -1,0 +1,9 @@
+-- | This module exports all container-related stuff.
+
+module Container
+       ( module Container.Class
+       , module Container.Reexport
+       ) where
+
+import Container.Class
+import Container.Reexport